Added for carbohydrateIntake multiple Food selection.

+++ Services/ChildGlucoCare.Services.Data/ICarbohydrateIntakeService.cs
@@ -5,10 +5,10 @@
     using ChildGlucoCare.Web.ViewModels.CarbohydtrateIntakes;
 
     public interface ICarbohydrateIntakeService
-    {
+    { 
         Task AddCarbohydrateIntakeAsync(AddNewCarbohydtrateIntakeViewModel carbohydrateIntake);
 
-        int GetBeuFromCarbohydrate();
+        double GetBeuFromCarbohydrate();
 
         string CalculateNeededInsulin(DateTime eatTime, double currentBloodGlucose, double insulinSensibility);
